Add environment variable, DMTCP_QUIET, and values of 0,1,2 (no -q, -q, -q -q), to suppress banner and JNOTE (-q) or banner/JNOTE/JWARNING (-q -q).

+++ test/gettimeofday.c
@@ -20,7 +20,7 @@ int main ()
     struct timeval tv;
     if (! getenv("NO_GETTIMEOFDAY"))
       gettimeofday(&tv, NULL);
-    printf("TIME (gettimeofday): %d\n",   tv.tv_sec);
+    printf("TIME (gettimeofday): %d\n", (int)tv.tv_sec);
     sleep(2);
   }
   return (0);
